Test-coding progressing
When running code that's supposed to utilize the
ds1302 chip/board combo, the output dates are (highly) off from where
they should be, plus they're static - no change as time is supposed to
be kicking over. Futhermore, this behavior happens whether the board is
wired onto the pico at the time of running or not. This tells me there
is something fatally flawed/damaged on this particular board and/or
chip (considering it *was* opperating as advertized, once-upon-a-time).
More chips and boards will need to be ordered ASAP in order to progress
with testing, code-writing, and documentating (not a word, ik; idc).