ds1302 lib modded
Despite the best intentions to make it more of a
universal function to have time-of-day and season included, but found
out that the built-in python rtc functions do not include such functions
and designing the code to expect such functions (by building them into
this module which might not get installed and/or utilized properly) sets
up the rest of the code to fail.
This is especially true since the most efficient use of the ds1302
assembly is to use it as the seed for the in-built RTC's time, with a
.txt file as the fail-over primary and manual-entry/default-value as the
fail-over secondaries. Because the in-built python RTC functions don't
include time of day or season, putting those functions in the ds1302.py
module would waste compute cycles and RAM space with unused data and
functions.

TL;DR: Due to fail-over design, the very thing I was trying to avoid
(time-of-day/season functions programed within individual project
applications) by putting them in ds1302.py became an actual necessity.
Ah well; good thought.